Compile with openssl 1.1.0
We moved @CPPUNIT_LIBS@ to later position.  This is because in my
development environment, it adds linker search directory which
contains system default openssl, while I'd like to use my custom
openssl installation.  This fix is not universal, since other
libraries can do the same thing.  It is a bit hard to fix this
properly because we mix -L and -l in a same variable.